backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
motor==3.7.1
authx==1.4.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
aiosmtplib==4.0.1
slowapi==0.1.9
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
httpx==0.28.1